fix(plugin): move options page to separate module and update registration
